fix: respect api import boundaries
